refactor: move logic in core module, refactor output formats
