+++ src/cli/commands/simple.rs
@@ -1,0 +1,9 @@
+use super::CommonArgs;
+use clap::Parser;
+
+/// Arguments for Simple mode
+#[derive(Parser, Debug)]
+pub struct SimpleArgs {
+    #[command(flatten)]
+    pub common: CommonArgs,
+}
