+++ src/core/sensor.rs
@@ -1,0 +1,8 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct Sensor {
+    pub name: String,
+    pub unit: String,
+    pub source: String,
+}
