+++ src/output/json.rs
@@ -1,14 +1,14 @@
 use std::fs::File;
 use std::io::Write;
 
-use anyhow::{Result, bail};
-use log::{info, trace};
+use anyhow::Result;
 use serde_json::json;
 
-use crate::config::{ListSensorsConfig, Mode, ProfileConfig};
-use crate::measurement::{MeasurementResult, PhaseMeasurementResult};
-use crate::output::{OutputFormatTrait, default_iterations_filename};
-use crate::source::Sensor;
+use crate::core::displayer::{
+    ListSensorsDisplayer, ProfilerDisplayer, default_iterations_filename,
+};
+use crate::core::measurement::{MeasurementResult, PhaseMeasurementResult};
+use crate::core::sensor::Sensor;
 use crate::util::file::{create_file_with_user_permissions, get_absolute_path};
 
 /// JSON output writer to file.
@@ -17,10 +17,10 @@ pub struct JsonOutput {
     filename: String,
 }
 
-impl OutputFormatTrait for JsonOutput {
-    fn simple_single(&mut self, config: &ProfileConfig, result: &MeasurementResult) -> Result<()> {
+impl ProfilerDisplayer for JsonOutput {
+    fn simple_single(&mut self, cmd: &[String], result: &MeasurementResult) -> Result<()> {
         let obj = json!({
-            "command": config.cmd.join(" "),
+            "command": cmd.join(" "),
             "mode": "simple",
             "metrics": result.metrics,
             "duration_ms": result.duration_ms,
@@ -32,18 +32,11 @@ impl OutputFormatTrait for JsonOutput {
         self.write_json(&obj)
     }
 
-    fn simple_iterations(
-        &mut self,
-        config: &ProfileConfig,
-        results: &[MeasurementResult],
-    ) -> Result<()> {
-        info!("Formatting {} simple iterations", results.len());
-
+    fn simple_iterations(&mut self, cmd: &[String], results: &[MeasurementResult]) -> Result<()> {
         let iters: Vec<_> = results
             .iter()
             .enumerate()
             .map(|(idx, result)| {
-                trace!("Formatting iteration {}", idx + 1);
                 json!({
                     "iteration": idx + 1,
                     "metrics": result.metrics,
@@ -56,7 +49,7 @@ impl OutputFormatTrait for JsonOutput {
             .collect();
 
         let root = json!({
-            "command": config.cmd.join(" "),
+            "command": cmd.join(" "),
             "mode": "simple-iterations",
             "iterations": iters
         });
@@ -66,19 +59,16 @@ impl OutputFormatTrait for JsonOutput {
 
     fn phases_single(
         &mut self,
-        config: &ProfileConfig,
+        cmd: &[String],
+        token_pattern: &str,
         result: &PhaseMeasurementResult,
     ) -> Result<()> {
         let phases_value = serde_json::to_value(result.phases.clone())?;
-        let phases_config = match &config.mode {
-            Mode::SimpleMode => bail!("Invalid configuration mode."),
-            Mode::PhaseMode(phases_config) => phases_config,
-        };
 
         let obj = json!({
-            "command": config.cmd.join(" "),
+            "command": cmd.join(" "),
             "mode": "phases",
-            "token_pattern": phases_config.token_pattern,
+            "token_pattern": token_pattern,
             "exit_code": result.exit_code,
             "phases": phases_value
         });
@@ -88,15 +78,10 @@ impl OutputFormatTrait for JsonOutput {
 
     fn phases_iterations(
         &mut self,
-        config: &ProfileConfig,
+        cmd: &[String],
+        token_pattern: &str,
         results: &[PhaseMeasurementResult],
     ) -> Result<()> {
-        info!("Formatting {} phase iterations", results.len());
-        let phases_config = match &config.mode {
-            Mode::SimpleMode => bail!("Invalid configuration mode."),
-            Mode::PhaseMode(phases_config) => phases_config,
-        };
-
         let iters: Vec<_> = results
             .iter()
             .enumerate()
@@ -111,16 +96,18 @@ impl OutputFormatTrait for JsonOutput {
             .collect();
 
         let root = json!({
-            "command": config.cmd.join(" "),
+            "command": cmd.join(" "),
             "mode": "phases-iterations",
-            "token_pattern": phases_config.token_pattern,
+            "token_pattern": token_pattern,
             "iterations": iters
         });
 
         self.write_json(&root)
     }
+}
 
-    fn list_sensors(&mut self, _config: &ListSensorsConfig, sensors: &[Sensor]) -> Result<()> {
+impl ListSensorsDisplayer for JsonOutput {
+    fn list_sensors(&mut self, sensors: &[Sensor]) -> Result<()> {
         self.write_json(&serde_json::to_value(sensors)?)
     }
 }
@@ -133,8 +120,6 @@ impl JsonOutput {
             .unwrap_or(default_iterations_filename("json"));
 
         let absolute_path = get_absolute_path(&filename)?;
-        info!("Creating JSON output file: {}", absolute_path);
-
         let file = create_file_with_user_permissions(&absolute_path)?;
 
         Ok(Self {
@@ -145,12 +130,8 @@ impl JsonOutput {
 
     fn write_json(&mut self, value: &serde_json::Value) -> Result<()> {
         let json_str = serde_json::to_string_pretty(value)?;
-        trace!("Writing JSON output ({} bytes)", json_str.len());
         writeln!(self.writer, "{}", json_str)?;
-
         println!("✔ JSON written to: {}", self.filename);
-        info!("JSON output saved to: {}", self.filename);
-
         Ok(())
     }
 }
