+++ src/util/mod.rs
@@ -1,2 +1,4 @@
 pub mod file;
+pub mod logging;
+pub mod platform;
 pub mod time;
